이미지 중복 처리 이전
