+++ src/main/java/com/halo/test/CKEditorAjax.java
@@ -7,15 +7,15 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/MZ")
-public class MZ extends HttpServlet {
+@WebServlet("/CKEditorAjax")
+public class CKEditorAjax extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 	}
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-
+		// 이곳은 CKEditor에 올라온 이미지를 ajax로 보내는 역할을 한다. 만든이 : 선생님
 		MZConfig.uploadFile(request, response);
 
 	}
