--- src/main/webapp/admin/boardmanagement/notice_test/0.js/testJW.js
@@ -1,26 +1,0 @@
-
-ClassicEditor
-	.create(document.querySelector('#classicNR'), {
-		simpleUpload: {
-			uploadUrl: 'http://localhost/halo/admin/0.js/ckeditor/file',
-
-			withCredentials: true,
-
-			// Headers sent along with the XMLHttpRequest to the upload server.
-			headers: {
-				'X-CSRF-TOKEN': 'CSRF-Token',
-				Authorization: 'Bearer <JSON Web Token>'
-			}
-		}
-
-
-	})
-	.then(editor => {
-		window.editor = editor;
-	})
-	.catch(error => {
-		console.error(error);
-	});
-
-
-
