--- src/main/webapp/admin/boardmanagement/notice_test/0.js/uploadAdpater2.js
@@ -1,14 +1,0 @@
-import UploadAdapter from '/js/ckeditor5/adapter/UploadAdapter.js'
-
-export default function makeEditor(target) {
-    return ClassicEditor.create(document.querySelector(target), {
-        extraPlugins: [MyCustomUploadAdapterPlugin]
-    })
-
-}
-
-function MyCustomUploadAdapterPlugin(editor) {
-    editor.plugins.get('FileRepository').createUploadAdapter = (loader) => {
-        return new UploadAdapter(loader)
-    }
-}
