+++ src/main/webapp/admin/boardmanagement/notice_test/noticeContent.jsp
@@ -7,7 +7,7 @@
 <meta charset="UTF-8" />
 <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 <title>notice</title>
-<script src="https://cdn.ckeditor.com/ckeditor5/40.2.0/classic/ckeditor.js"></script>
+<script src="admin/_js/ckeditor/build/ckeditor.js"></script>
 <script src="admin/_js/test2.js"></script>
